@@ -19,9 +19,8 @@ package org.apache.shardingsphere.elasticjob.lite.console.restful;
 
 import org.apache.shardingsphere.elasticjob.lite.console.service.JobAPIService;
 import org.apache.shardingsphere.elasticjob.lite.console.service.impl.JobAPIServiceImpl;
-import org.apache.shardingsphere.elasticjob.lite.lifecycle.domain.ShardingInfo;
 import org.apache.shardingsphere.elasticjob.lite.lifecycle.domain.JobBriefInfo;
-import com.google.common.base.Optional;
+import org.apache.shardingsphere.elasticjob.lite.lifecycle.domain.ShardingInfo;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -71,7 +70,7 @@ public final class JobOperationRESTfulAPI {
     @POST
     @Path("/{jobName}/trigger")
     public void triggerJob(@PathParam("jobName") final String jobName) {
-        jobAPIService.getJobOperatorAPI().trigger(Optional.of(jobName), Optional.<String>absent());
+        jobAPIService.getJobOperatorAPI().trigger(jobName);
     }
     
     /**
@@ -83,7 +82,7 @@ public final class JobOperationRESTfulAPI {
     @Path("/{jobName}/disable")
     @Consumes(MediaType.APPLICATION_JSON)
     public void disableJob(@PathParam("jobName") final String jobName) {
-        jobAPIService.getJobOperatorAPI().disable(Optional.of(jobName), Optional.<String>absent());
+        jobAPIService.getJobOperatorAPI().disable(jobName, null);
     }
     
     /**
@@ -95,7 +94,7 @@ public final class JobOperationRESTfulAPI {
     @Path("/{jobName}/disable")
     @Consumes(MediaType.APPLICATION_JSON)
     public void enableJob(@PathParam("jobName") final String jobName) {
-        jobAPIService.getJobOperatorAPI().enable(Optional.of(jobName), Optional.<String>absent());
+        jobAPIService.getJobOperatorAPI().enable(jobName, null);
     }
     
     /**
@@ -107,7 +106,7 @@ public final class JobOperationRESTfulAPI {
     @Path("/{jobName}/shutdown")
     @Consumes(MediaType.APPLICATION_JSON)
     public void shutdownJob(@PathParam("jobName") final String jobName) {
-        jobAPIService.getJobOperatorAPI().shutdown(Optional.of(jobName), Optional.<String>absent());
+        jobAPIService.getJobOperatorAPI().shutdown(jobName, null);
     }
     
     /**
