@@ -17,12 +17,10 @@
 
 package org.apache.shardingsphere.elasticjob.lite.internal.util;
 
-import com.google.common.base.Function;
 import com.google.common.base.Joiner;
-import com.google.common.collect.Lists;
-import org.apache.shardingsphere.elasticjob.lite.util.env.IpUtils;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+import org.apache.shardingsphere.elasticjob.lite.util.env.IpUtils;
 
 import java.util.HashMap;
 import java.util.List;
@@ -30,6 +28,7 @@ import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 /**
  * Sensitive info utility.
@@ -48,25 +47,21 @@ public final class SensitiveInfoUtils {
     public static List<String> filterSensitiveIps(final List<String> target) {
         final Map<String, String> fakeIpMap = new HashMap<>();
         final AtomicInteger step = new AtomicInteger();
-        return Lists.transform(target, new Function<String, String>() {
-            
-            @Override
-            public String apply(final String input) {
-                Matcher matcher = Pattern.compile(IpUtils.IP_REGEX).matcher(input);
-                String result = input;
-                while (matcher.find()) {
-                    String realIp = matcher.group();
-                    String fakeIp;
-                    if (fakeIpMap.containsKey(realIp)) {
-                        fakeIp = fakeIpMap.get(realIp);
-                    } else {
-                        fakeIp = Joiner.on("").join(FAKE_IP_SAMPLE, step.incrementAndGet());
-                        fakeIpMap.put(realIp, fakeIp);
-                    }
-                    result = result.replace(realIp, fakeIp);
+        return target.stream().map(input -> {
+            Matcher matcher = Pattern.compile(IpUtils.IP_REGEX).matcher(input);
+            String result = input;
+            while (matcher.find()) {
+                String realIp = matcher.group();
+                String fakeIp;
+                if (fakeIpMap.containsKey(realIp)) {
+                    fakeIp = fakeIpMap.get(realIp);
+                } else {
+                    fakeIp = Joiner.on("").join(FAKE_IP_SAMPLE, step.incrementAndGet());
+                    fakeIpMap.put(realIp, fakeIp);
                 }
-                return result;
+                result = result.replace(realIp, fakeIp);
             }
-        });
+            return result;
+        }).collect(Collectors.toList());
     }
 }
