@@ -166,13 +166,7 @@ public final class ZookeeperRegistryCenter implements CoordinatorRegistryCenter 
     public List<String> getChildrenKeys(final String key) {
         try {
             List<String> result = client.getChildren().forPath(key);
-            Collections.sort(result, new Comparator<String>() {
-                
-                @Override
-                public int compare(final String o1, final String o2) {
-                    return o2.compareTo(o1);
-                }
-            });
+            result.sort(Comparator.reverseOrder());
             return result;
          //CHECKSTYLE:OFF
         } catch (final Exception ex) {
