@@ -38,7 +38,7 @@ public final class RotateServerByNameJobShardingStrategyTest {
         Map<JobInstance, List<Integer>> expected = new HashMap<>();
         expected.put(new JobInstance("host1@-@0"), Collections.singletonList(0));
         expected.put(new JobInstance("host2@-@0"), Collections.singletonList(1));
-        expected.put(new JobInstance("host0@-@0"), Collections.<Integer>emptyList());
+        expected.put(new JobInstance("host0@-@0"), Collections.emptyList());
         assertThat(rotateServerByNameJobShardingStrategy.sharding(Arrays.asList(new JobInstance("host0@-@0"), new JobInstance("host1@-@0"), new JobInstance("host2@-@0")), "1", 2), is(expected));
     }
     
@@ -47,7 +47,7 @@ public final class RotateServerByNameJobShardingStrategyTest {
         Map<JobInstance, List<Integer>> expected = new HashMap<>();
         expected.put(new JobInstance("host2@-@0"), Collections.singletonList(0));
         expected.put(new JobInstance("host0@-@0"), Collections.singletonList(1));
-        expected.put(new JobInstance("host1@-@0"), Collections.<Integer>emptyList());
+        expected.put(new JobInstance("host1@-@0"), Collections.emptyList());
         assertThat(rotateServerByNameJobShardingStrategy.sharding(Arrays.asList(new JobInstance("host0@-@0"), new JobInstance("host1@-@0"), new JobInstance("host2@-@0")), "2", 2), is(expected));
     }
     
@@ -56,7 +56,7 @@ public final class RotateServerByNameJobShardingStrategyTest {
         Map<JobInstance, List<Integer>> expected = new HashMap<>();
         expected.put(new JobInstance("host0@-@0"), Collections.singletonList(0));
         expected.put(new JobInstance("host1@-@0"), Collections.singletonList(1));
-        expected.put(new JobInstance("host2@-@0"), Collections.<Integer>emptyList());
+        expected.put(new JobInstance("host2@-@0"), Collections.emptyList());
         assertThat(rotateServerByNameJobShardingStrategy.sharding(Arrays.asList(new JobInstance("host0@-@0"), new JobInstance("host1@-@0"), new JobInstance("host2@-@0")), "3", 2), is(expected));
     }
 }
