@@ -56,17 +56,13 @@ public final class EmbedTestingServer {
             // CHECKSTYLE:ON
             RegExceptionHandler.handleException(ex);
         } finally {
-            Runtime.getRuntime().addShutdownHook(new Thread() {
-                
-                @Override
-                public void run() {
-                    try {
-                        testingServer.close();
-                    } catch (final IOException ex) {
-                        RegExceptionHandler.handleException(ex);
-                    }
+            Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+                try {
+                    testingServer.close();
+                } catch (final IOException ex) {
+                    RegExceptionHandler.handleException(ex);
                 }
-            });
+            }));
         }
     }
 }
