@@ -29,6 +29,6 @@ public final class HostExceptionTest {
     @Test
     public void assertGetCause() {
         IOException cause = new IOException();
-        assertThat(new HostException(cause).getCause(), Is.<Throwable>is(cause));
+        assertThat(new HostException(cause).getCause(), Is.is(cause));
     }
 }
