@@ -17,7 +17,6 @@
 
 package org.apache.shardingsphere.elasticjob.lite.lifecycle.api;
 
-import com.google.common.base.Optional;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.elasticjob.lite.lifecycle.internal.operate.JobOperateAPIImpl;
@@ -42,7 +41,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job settings API
      */
-    public static JobSettingsAPI createJobSettingsAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static JobSettingsAPI createJobSettingsAPI(final String connectString, final String namespace, final String digest) {
         return new JobSettingsAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
     
@@ -54,7 +53,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job operate API
      */
-    public static JobOperateAPI createJobOperateAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static JobOperateAPI createJobOperateAPI(final String connectString, final String namespace, final String digest) {
         return new JobOperateAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
     
@@ -66,7 +65,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job sharding operate API
      */
-    public static ShardingOperateAPI createShardingOperateAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static ShardingOperateAPI createShardingOperateAPI(final String connectString, final String namespace, final String digest) {
         return new ShardingOperateAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
     
@@ -78,7 +77,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job statistics API
      */
-    public static JobStatisticsAPI createJobStatisticsAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static JobStatisticsAPI createJobStatisticsAPI(final String connectString, final String namespace, final String digest) {
         return new JobStatisticsAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
     
@@ -90,7 +89,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job server statistics API
      */
-    public static ServerStatisticsAPI createServerStatisticsAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static ServerStatisticsAPI createServerStatisticsAPI(final String connectString, final String namespace, final String digest) {
         return new ServerStatisticsAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
     
@@ -102,7 +101,7 @@ public final class JobAPIFactory {
      * @param digest registry center digest
      * @return job sharding statistics API
      */
-    public static ShardingStatisticsAPI createShardingStatisticsAPI(final String connectString, final String namespace, final Optional<String> digest) {
+    public static ShardingStatisticsAPI createShardingStatisticsAPI(final String connectString, final String namespace, final String digest) {
         return new ShardingStatisticsAPIImpl(RegistryCenterFactory.createCoordinatorRegistryCenter(connectString, namespace, digest));
     }
 }
